cleanup!: Remove poyo-specific `torch_brain.utils`
This PR removes three utils in `torch_brain.utils` that are present just
for POYO:

- create_linspace_latent_tokens
- create_start_end_unit_tokens
- resolve_weights_based_on_interval_membership

These should not be presented as regular public API in
`torch_brain.utils`, but only as POYO-specific API. These have been
moved to appropriate locations in the code.

Closes #291
