+++ torch_brain/utils/__init__.py
@@ -1,8 +1,7 @@
 from .binning import bin_spikes
 from .misc import calculate_sampling_rate as calculate_sampling_rate
 from .misc import np_string_prefix
-from .tokenizers import create_linspace_latent_tokens, create_start_end_unit_tokens
-from .weights import isin_interval, resolve_weights_based_on_interval_membership
+from .weights import isin_interval
 
 # from .stitcher import stitch
 
@@ -16,9 +15,6 @@ def seed_everything(*args, **kwargs):
 
 __all__ = [
     # "stitch",
-    "create_linspace_latent_tokens",
-    "create_start_end_unit_tokens",
-    "resolve_weights_based_on_interval_membership",
     "isin_interval",
     "np_string_prefix",
     "bin_spikes",
